python-backend-serverless/requirements.txt
openai-agents
pydantic
fastapi
uvicorn
redis
python-dotenv
